Check if file exists

+++ install/delta.sh
@@ -1,2 +1,4 @@
-mv ~/.gitconfig{,.bak}
+if [ -e .gitconfig ]; then
+	mv ~/.gitconfig{,.bak}
+fi
 cp ~/.local/share/${project}/configs/.gitconfig ~/.gitconfig
